removed some unused code in high level components. added feedback form file
