+++ src/App.js
@@ -3,7 +3,7 @@ import { Home, About } from "./pages/Other/";
 import { SeasonSchedule, TodaysSchedule } from "./pages/Scheduling/index.js";
 import GameInfo from "./pages/GameInfo/GameInfo";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Navigation from "./components/Navigation/Navigation";
+import { FeedbackForm, Navigation } from "./components/";
 
 // Not sure if this is best practice but I use App.js essentially as a router
 function App() {
@@ -11,6 +11,7 @@ function App() {
     <Router>
       <div className="App">
         <Navigation />
+        <FeedbackForm />
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/todays-schedule" element={<TodaysSchedule />} />
