+++ src/components/FeedbackForm.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function FeedbackForm() {
+  return (
+    <span>Report an issue</span>
+  )
+}
