+++ src/components/index.js
@@ -1,0 +1,4 @@
+import FeedbackForm from "./FeedbackForm";
+import Navigation from "./Navigation";
+
+export { FeedbackForm, Navigation };
