+++ src/pages/GameInfo/components/scorecards/ScoringRow5.jsx
@@ -1,5 +1,3 @@
-// Legacy code migrated to ScorecardRow, deletion pending
-
 import React from "react";
 import ScoringCell5 from "./ScoringCell5";
 import EmptyCell from "./EmptyCell";
@@ -18,3 +16,50 @@ export default function ScoringRow5({ order }) {
     </tr>
   );
 }
+
+/* from ScorecardRowDion.jsx 
+
+import React from "react";
+import ScoringCell from "../../CombinedScorecard/components/ScoringCell";
+import EmptyCell from "../../CombinedScorecard/components/EmptyCell";
+import { defensivePositionMapper } from "../../../util/constants";
+
+export default function ScorecardRowDion({ order, players, innings, idx }) {
+  return (
+    <tr>
+      {/* <th>14</th>
+      <th>Dion </th>
+      // <th>Leung </th> }
+      <th>
+        {players.map(player => {
+          return <p>{player.jerseyNumber}</p>;
+        })}
+      </th>
+      <th>
+        {players.map(player => {
+          return (
+            <p>
+              {player.preferredName} {player.lastName}
+            </p>
+          );
+        })}
+      </th>
+      <th>
+        {players.map(player => {
+          return (
+            <p>{defensivePositionMapper[player.positionNumber].positionAbbr}</p>
+          );
+        })}
+      </th>
+      {innings.map(inning => {
+          if (inning.involved === true) {
+            return <ScoringCell scoringEvent={inning.scoringOutcome}/>;
+          } else {
+            return <EmptyCell />;
+          }
+      })}
+    </tr>
+  );
+}
+
+*/
